The logic of searching for discounts has been created, the user info has been redesigned and the rewrite card has been changed.

+++ routes/categories.api.js
@@ -7,7 +7,7 @@ const router = Router()
 router.post(
   '/add',
   [
-    check('name', 'name isEmpty').isLength({min:2})
+    check('name', 'name isEmpty').isLength({ min: 2 })
   ],
   async (req, res) => {
     try {
@@ -20,13 +20,13 @@ router.post(
         })
       }
 
-      const {name} = req.body
-      const candidate = await Category.findOne({name})
-      if(candidate) {
+      const { name } = req.body
+      const candidate = await Category.findOne({ name })
+      if (candidate) {
         return res.status(400).json({ message: "Такая категория уже есть", status: false })
       }
 
-      const category = new Category({name})
+      const category = new Category({ name })
       await category.save()
       const categories = await Category.find()
       res.status(201).json({ message: "Новая категория создана", status: true, categories })
@@ -53,10 +53,9 @@ router.get(
   '/get/client',
   async (req, res) => {
     try {
-      const categories = await Category.find({show: true})
-      res.json({categories, status: true})
+      const categories = await Category.find({ show: true })
+      res.json({ categories, status: true })
     } catch (e) {
-      console.log(e)
       res.status(500).json({ message: "Что-то пошло не так, перезагрузите страницу" })
     }
   }
@@ -66,13 +65,15 @@ router.post(
   '/del',
   async (req, res) => {
     const { id } = req.body
-    console.log(id)
     try {
+      const category = await Category.findOne({ _id: id })
+      if (category.name === 'Все') {
+        return res.json({ message: "Эту категорию нельзя удалить. (с) Администратор", status: true, categories })
+      }
       await Category.findByIdAndDelete({ _id: id })
       const categories = await Category.find()
       res.json({ message: "Категория была удалена", status: true, categories })
     } catch (e) {
-      console.log(e)
       res.status(500).json({ message: "Что-то пошло не так, перезагрузите страницу" })
     }
   }
@@ -84,9 +85,9 @@ router.post(
     const { id, checked } = req.body
     try {
       await Category.findByIdAndUpdate(
-        {_id: id},
-        {$set: {'show': checked}}
-        )
+        { _id: id },
+        { $set: { 'show': checked } }
+      )
       const categories = await Category.find()
       res.json({ message: "Обновлено", status: true, categories })
     } catch (e) {
